# omcgo/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OMC API
  description: Operations Management and Control for Base Stations
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/datamodels:
    get:
      tags: [DataModels]
      summary: List data models
      operationId: listDataModels
      parameters:
        - $ref: "#/components/parameters/CarrierParam"
        - $ref: "#/components/parameters/TechnologyParam"
        - name: oui
          in: query
          schema:
            type: string
        - name: product_class
          in: query
          schema:
            type: string
        - name: scope
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated data model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/DataModel"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [DataModels]
      summary: Create a data model
      operationId: createDataModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDataModelRequest"
      responses:
        "201":
          description: Data model created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataModel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/datamodels/resolve:
    get:
      tags: [DataModels]
      summary: Resolve data model by OUI, product class, and carrier
      operationId: resolveDataModel
      parameters:
        - name: oui
          in: query
          schema:
            type: string
        - name: product_class
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/CarrierParam"
      responses:
        "200":
          description: Resolved data model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataModel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/datamodels/statistics:
    get:
      tags: [DataModels]
      summary: Get data model statistics
      operationId: getDataModelStatistics
      responses:
        "200":
          description: Data model statistics
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/datamodels/import:
    post:
      tags: [DataModels]
      summary: Bulk import data models
      operationId: importDataModels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Import completed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/datamodels/cache/refresh:
    post:
      tags: [DataModels]
      summary: Refresh data model cache
      operationId: refreshDataModelCache
      responses:
        "200":
          description: Cache refreshed
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/datamodels/{id}:
    get:
      tags: [DataModels]
      summary: Get data model by ID
      operationId: getDataModel
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "200":
          description: Data model details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataModel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [DataModels]
      summary: Update a data model
      operationId: updateDataModel
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDataModelRequest"
      responses:
        "200":
          description: Data model updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataModel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [DataModels]
      summary: Delete a data model
      operationId: deleteDataModel
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "204":
          description: Data model deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/datamodels/{id}/export:
    get:
      tags: [DataModels]
      summary: Export a data model
      operationId: exportDataModel
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "200":
          description: Data model export
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/datamodels/{id}/activate:
    post:
      tags: [DataModels]
      summary: Activate a data model
      operationId: activateDataModel
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "200":
          description: Data model activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataModel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/datamodels/{id}/deprecate:
    post:
      tags: [DataModels]
      summary: Deprecate a data model
      operationId: deprecateDataModel
      parameters:
        - $ref: "#/components/parameters/UUIDPathParam"
      responses:
        "200":
          description: Data model deprecated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataModel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===========================================================================
  # OUI
  # ===========================================================================
  /api/v1/oui:
    get:
      tags: [OUI]
      summary: List OUIs
      operationId: listOUIs
      responses:
        "200":
          description: OUI list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OUI"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [OUI]
      summary: Create an OUI
      operationId: createOUI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOUIRequest"
      responses:
        "201":
          description: OUI created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OUI"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ===========================================================================
  # Templates
  # ===========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    UUIDPathParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CarrierParam:
      name: carrier
      in: query
      description: Carrier / operator name
      schema:
        type: string
    TechnologyParam:
      name: technology
      in: query
      description: Radio access technology (e.g. 4G, 5G)
      schema:
        type: string

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"
        details:
          type: string
          description: Additional error details
        request_id:
          type: string
          description: Request tracking ID
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          description: Total number of records
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 20

    # ---- Auth ----
...
    DataModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        carrier:
          type: string
        technology:
          type: string
        oui:
          type: string
        product_class:
          type: string
        scope:
          type: string
        status:
          type: string
          enum: [draft, active, deprecated]
        parameters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              writable:
                type: boolean
              description:
                type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateDataModelRequest:
      type: object
      required: [name, version]
      properties:
        name:
          type: string
        version:
          type: string
        carrier:
          type: string
        technology:
          type: string
        oui:
          type: string
        product_class:
          type: string
        scope:
          type: string
        parameters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              writable:
                type: boolean
              description:
                type: string
    UpdateDataModelRequest:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        carrier:
          type: string
        technology:
          type: string
        scope:
          type: string
        parameters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              writable:
                type: boolean
              description:
                type: string

    # ---- OUI ----
    OUI:
      type: object
      properties:
        id:
          type: string
          format: uuid
        oui:
          type: string
          example: "00259E"
        manufacturer:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
    CreateOUIRequest:
      type: object
      required: [oui, manufacturer]
      properties:
        oui:
          type: string
        manufacturer:
          type: string
        description:
          type: string

    # ---- ConfigTemplate ----
...